Allow for use of new workflow
This allows the widget to use the new workflow, as implemented here - https://git.dev.d2l/projects/CORE/repos/lp/pull-requests/5292/overview. The real difference is that we do away with the adding/removing filters on the HM API, and instead have a direct link to the semesters and departments. In addition, the results returned by these APIs are more correct, as they return relevant results, rather than directly-enrolled results.