Actually delay loading of filter menu
Thought this was already the case, but apparently not. We want to delay the loading of the menu contents until it's opened for the first time. Also, hide the loading spinner at approximately the right time, rather than right away (not entirely the correct time, since the d2l-list-item-filter's still have to do their organization requests, but that's harder to keep track of).