Refactor d2l-course-management-behavior
The vast majority of the functionality contained within this behavior was either only used in one of d2l-my-courses/d2l-all-courses, or it was overridden in one or both anyway. So, moved things up a level to make it clearer (aside from the few things that actually _are_ common to both).