DE27525 Fix unpinning enrollment in main screen
This issue was actually two problems - first, the console error was happening as we changed to the enrolledUser HM API a while back, but that broke the extraction of the org unit ID here. Switched to use the organization instead.

Second, the animation was failing to complete. This I cannot explain, as there was never a handler for this event here as far as I can recall, but adding the handler does fix it. Not sure how it was working before...