Use d2l-fetch from bower, to take advantage of version resolution
Under this new method, `.use`ing the various middlewares is done by the consumer of the element, in this case BSI. This means we can get the same versions for everything, and the same middlewares. In addition, My Courses shouldn't really _care_ how the d2l-fetch sausage is made, only that it gets made somehow.

Unclear whether I really need to have the HTML import at all in My Courses, but hopefully if nothing else it makes it more obvious that it's not just an unused bower reference.