Switch to directly setting searchUrl
The role filter tab works distinctly from the existing department/semester filters, which resulted in me realizing that the search isn't really doing state transfer correctly. A better approach is for the search widget to simply take an arbitrary search URL, query it, and return the results. This URL can be modified by search text, sort order, or filters; when the results come back, we update the search URL to be the self URL of the response, which allows us to again modify the search URL -> get results -> update search URL from response.

This more correctly transfers state between the client and server, and means that at any given moment, the self URL of the enrollments search result is representative of what the user is seeing. This also allows for the role filter tab to work more naturally, as the process for filtering by roles is a bit more complex - but with this new approach, it will work the same way as everything else, i.e. just update the search URL (then the search widget fetches that URL, and returns the results).