US90278 Ditch currentFilters and be a bit nicer
The currentFilters property was a) a lie, and b) not very nice. It was a lie as it only contained the semester and department filters, not role filters. It was not very nice because it meant d2l-all-courses was kinda reaching inside the filter menu to get the opener text, whereas really we can just push that number upwards to it.

Also fixed the Role tab naming and clearing, which wasn't yet implemented.