US90278 Fix clearing role filters issue
When we clear the filters, reset the state by setting the `include` parameter to be blank, and re-fetch the roles. Ideally, this would be on the /role-filters entity as an action, but it hasn't been added yet. Once it is, this can be modified to just use that action rather than explicitly setting `include: ''` in the URL creation.