Clean up hiding things
Yeeeaaah, this is kinda out of nowhere, but didn't want to to inflate another PR. This just leverages the `hidden` attribute more (and computed properties, where possible), and explicitly `toggleClass`-ing less. I personally find this method a lot easier to read, as it includes any hide/unhide stuff in the DOM, which is always nice. Thoughts/opinions?