Re-set myEnrollmentsEntity on search results
This pushes the change to the entity down via bindings. In general, this is "the right way" to do this, and it solves the issue where using a role filter and a department/semester in conjunction didn't work as expected as the /role-filters entity wasn't getting updated with the new `?via` value whenever we selected a department/semester (i.e. we weren't getting the new `parentOrganizations` in the `via` URL).