DE27982 Combine role filters with the same name
This isn't a super performant approach, although it is probably the most robust this can be approached (that I can think of). The idea is that we combine role filters with the same name (easy enough), and when a multi-filter item is selected, we make subsequent "add-filter" action calls for each of the individual filters. This needs to be sequentially, unfortunately - doing them in parallel would just end up in a state where only one is enabled - but it's probably "fast enough", especially once things are cached. If we want to hide the loading, it might also be a good idea to disable further input in the filter menu until the result set is actually updated.