DE27982 New method for multi-filters
This is a simpler method of enabling/disabling filters which have more than one filter of the same title. Instead of throwing things around in events, we instead just fire the title, and the handler will continue to enable filters by that title which are off until there are none left - at which point, it fires the role-filters-changed event like before.