US91774 Replace delayLoad with IntersectionObserver
Effectively, delayLoad was a method of preventing My Courses from loading information for courses that were only in the overlay on initial load (i.e. unpinned courses). IntersectionObserver offers a cleaner way to do this - a course tile will only load itself when it becomes visible, which is only when the overlay is opened for those unpinned courses.