Remove IE11 from Sauce
Try as I might, I simply could not get these working in IE11 - three of the tests (of 220!) would consistently fail with extremely unhelpful error messages. Sigh :( But, the value of being able to run the Sauce tests in all the other browsers remains, and is certainly better than not running it at all, which is what we were doing before.