Re-enable troublesome tests
These tests were failing on Sauce on OSX browsers - sometimes Chrome, sometimes Firefox, sometimes Safari. Not entirely sure why, but this change was initially just intended to move the getFutureDate, getPastDate, and verifyOverlay functions into the inner test scope. With a bit of further poking, it seems that the root of the problem was the Date.now() happening only once, and the future/end dates being only 8 seconds off that. One method that worked was making the future/end dates 8,000 seconds off the current date, but for an even cleaner approach, we can re-make the future/end date for each test.