US91774 Fix Chrome/FF behaviour
Weirdly, Chrome and Firefox will both call the IntersectionObserver callback when you initially set it up with IntersectionObserver.observe(). This is documented and is an intentional recent change, but is still somewhat strange... either way, way to get around it is to just check and make sure we have a non-zero IntersectionObserverEntry.intersectionRatio in the callback, meaning the tile is at least partially visible, before populating its data.