US91774 Improve performance measures
Adds the `d2l.my-courses.visible-images` measure, which is an indication of when the visible tiles have finished loading (i.e. their organizations and images have been fetched). This is a naive implementation, as it relies on the first widgets to be complete _being_ the visible ones (generally true, but not guaranteed), and relies on all course tiles (or at least, the visible ones) being attached before the first /organizations request comes back (again, generally true but not guaranteed).