US91774 Rework performance marks and measures
This change restructures things a bit, so that the loading measures are all fairly human-centric. In particular,

- d2l.my-courses.meaningful.visible: time until all visible organizations are complete; this marks when the widget is functional for a sighted user (includes course names, and tiles are clickable, doesn't include course images)
- d2l.my-courses.meaningful.all: time until all organizations are complete; this marks when the widget is functional for a screen reader (same requirements as above, but for all courses)
- d2l.my-courses: this has changed, now marks when widget is complete for a sighted user (includes course names and images, but only for visible tiles)

The "missing" measurement here is time until the last course image is loaded; however, this will be ignored as the next change is to only ever load an image when it becomes visible (since they're not required for a screenreader anyway).