Listen for set-course-image events from anywhere
This only applies to the case where you have a course banner on a homepage, and also have the My Courses widget on that homepage, but might as well. This means when you change the image in the banner, the course tile's image will also update. (The opposite direction, changing image from course tile -> update image in banner, is already implemented in d2l-image-banner-overlay.)