DE28442 Fetch all pinned courses in widget
This was... probably supposed to always be here? Either way, this is needed functionality for US92691, so two birds one stone - continue to fetch subsequent pages of enrollments until we hit a course that is unpinned, so that a user with >25 pinned courses will still see all of their pinned courses in the widget.

Also removed the _enrollmentsSearchUrl property, as it can just be a local variable to the _fetchEnrollments function.