Remove unpinnedEnrollments list
With the decoupling of d2l-all-courses and d2l-my-courses, the latter no longer needs to keep track of unpinned enrollments. It still needs to _remove_ enrollments when they get unpinned (either from d2l-all-courses or the course selector), but doesn't need to hang on to them.

Also removed _lastEnrollmentsSearchResponse, as it was unused.