Remove d2l-course-management-behavior from d2l-my-courses
This added the pinnedEnrollments and unpinnedEnrollments properties, as well as the _setEnrollmentPinData function. The function was being used once, but has no real effect, since we're no longer moving items from one array to another, but rather adding or removing them from a single array. With unpinnedEnrollments now being removed as well, we can just use a local/private _pinnedEnrollments array in place of the behavior-defined public pinnedEnrollments one now.