US92691 Introduce thin main component
Right now, this just loads d2l-my-courses-content-animated. In the future, it will use the updatedSortLogic parameter to load an entirely different d2l-my-courses-content component, which will support the 12-course logic for the US, as well as other changes that we only want to exist in the new version of the widget.