US92691 Add new d2l-my-courses-content
This new content component accomplishes the goal of the US, which was to fill up to 12 enrollments for a user. In particular, we will now fetch all of a user's pinned enrollments, then continue to add enrollments until we hit 12, or they no longer have any enrollments. There are other "new" things in here as well, for example using the new query parameters `?sort=current&promotePins=true`. We also no longer auto-pin in the new logic, so we explicitly set it to false to be safe for the initial request.

Aside from those changes, the component itself is largely the same as the older one, hence breaking out the behaviour in a previous PR.

The idea behind the structure here is that when the `updatedSortLogic` variable is removed, we can just remove `d2l-my-courses-content-animated`, and optionally re-integrate the functionality in `d2l-my-courses-content-behavior` right back into `d2l-my-courses-content` itself.