Add tests for d2l-my-courses-content
Mostly based off of existing tests of `d2l-my-courses-content-animated` (which themselves are based off the old tests for the original `d2l-my-courses`), updated as required. Also found a couple small... ehhhh, "bugs" is too strong of a word, but a couple things that could be a little cleaner in the non-test code that were found in writing said tests.