DE28618 Hide past courses
To prevent a user from just seeing a wall of ended courses (as they will show up after future courses in Current sort), hide them from the main widget view. In doing so with CSS, I realized that we can take a similar CSS-based approach to limiting the widget to 12 items - which removes the need to keep two arrays up to date, so _allEnrollments was removed. Now, the `limit-to-12` attribute on the `d2l-course-tile-grid` does the work for us.

Note that this only applies for users using Current sorting (users with large numbers of enrollments will be using LastAccessed, and will see past courses in the widget), and the "hide if ended" rule is always overridden by the enrollment being pinned.

Also slightly reworked when we show the widget - we set _showContent = true now either after the last visible organization has loaded, or at worst one second after the first visible organization has loaded. This is to prevent some jank as things are loading, and generally makes the experience a bit more polished.