Fix some small pinning issues
First issue: pinning an unpinned filled-to-12 enrollment would cause the enrollment to show up twice. This was because we would add the enrollment to the beginning of the list of enrollments, but not remove the old copy of it below the pins. Simple enough to fix, we just need to always remove the enrollment on a `d2l-course-pinned-change` event, then in the case that it's being pinned, we need to re-add it.

Second issue: related to the first, when an enrollment was pinned and added to the beginning of the list, it wouldn't show up as pinned. However, we do have the self link of the enrollment, so we can use that to re-fetch the enrollment (which will have the correct pinned status on the response). Bonus to this is that it animates quite nicely!