DE28649 Add newly-pinned enrollments to end of pinned enrollments
Previously in My Courses, we wanted the most-recently-pinned course first, but with the updated sort order changes, we want it to be the opposite. This is going back to how it was before, which is also consistent with the course selector. This change fixes My Courses so that when a course is pinned in the course selector, it moves to the end of the already-pinned items in My Courses. When the same is done to unpin something, it moves to just-below the pinned items.

Note that this doesn't change the behavior for pinning/unpinning a course tile - in that case, the tile does not move, regardless of its current position. Refreshing the page will result in things being in order in this case, as moving them around when somebody is interacting with the widget directly is pretty janky and weird.