DE28675 Fix "no unpinned" message when config is off
Accidentally broke this when the config is off and the user has only unpinned courses. Previously _showContent = true was only happening in the handler when organization requests had finished, but this wasn't working when the config was off (since none of said organizations would be visible, as they're unpinned). Moved the handler to be a bit more like it was before, but still with the worst-case-one-second setTimeout. In other words, the content should become visible either after the last visible organization has loaded, or one second after the enrollments request has finished.