DE28697 Fire d2l-pinned-course-change from View All Courses
This allows the main widget to update its list of courses when a user pins or unpins something in the View All Courses overlay. The good news is that in this case, we can pass the updated enrollment directly via the event, which means we don't need to do the org unit ID fiddling. #nice.