US93378 Change pageSize to 20
Ultimately we were only fetching the larger page size so that we could render a "View All Courses (50+)" link - unfortunately, the response time of this request is pretty much linearly related to page size. This means a page size of 50 takes about twice as long as a page size of 20. On fast connections, not a particularly noticeable difference, but on crappy 3G in a tunnel through a mountain, this can mean loading in 10 seconds rather than 5 (not great either way, but y'know). The only downside of this change is that the View All Courses link now reads "(20+)", but ultimately we're not using the "99+" convention either way, so the actual number doesn't really mean all that much.

Also added round-down-to-10 logic, otherwise it's possible to end up with things like "View All Courses (23+)" if you pin/unpin a few things.