DE28699 Include past courses in limit-to-12 count (sometimes)
Bit of an oversight here - when a user has many enrollments, they are sorted by last accessed date. This means there's no `hide-past-courses` being applied to the course tile grid, which is correct. However, we were missing the CSS rule for including these (visible) past courses in the `limit-to-12` CSS logic. This meant that all recently-accessed ended courses were not contributing to the 12 limit, so the user ended up seeing more than 12.