Re-enable some tests
These tests were added in #455, but intentionally skipped as they were failing in Safari on Sauce. The root cause here appeared to be missing organizations links, which goes unnoticed on other browsers... weird. Fixed the tests, and added a couple safety checks in `d2l-course-tile` to prevent these sorts of (unlikely) errors from happening in the wild. This fixes the tests on a local Safari instance, but unfortunately _some_ of the tests are still problematic on Sauce. Going to skip them for now, so that at least we can get the majority of these tests running.