First stage of new d2l-course-image-tile
This version includes most of the basic functionality that d2l-course-tile has, but excludes functionality around course updates (fetching/displaying), pin/unpin (actually changing the state, that is), inactive/ended/not-yet-started overlays, started-inactive-pinned notification and alert, and setting/uploading course images. These will come in stages over the next PR(s).

Currently, you must set the `css-grid-view` attribute on `d2l-my-courses` itself in order for this new component to be used in place of the old one. (Upcoming LP PR for this, too.)