US90524 IE11 support
This allows for the CSS grid layout to work in IE11. Unfortunately IE11 uses an old CSS grid spec, so we only use this when required, but it works, and is ignored by all other browsers. This does include a bit of a hack, but only in IE - sometimes it takes a moment for the _enrollments array to get populated, and therefore there are no course tiles to select. In this case, we simply re-_onResize a few times (for up to a second) until things get properly rendered. (Tested this fairly extensively - it only occurs once in a while, but the provided solution worked in all cases.)

This change did require ditching the `grid-gap`, in favour of `grid-gap-column` and `margin-bottom` on the course tiles. This is because the method by which below-the-fold tiles are hidden has changed slightly. If we continue to use grid-gap, it adds space between the last visible course tile and the View All Course link, since the tile-surrounding divs are still technically visible (albeit empty). The new method is pretty much visually identical, and prevents the extra space.

Two other small changes:

- Increased the tile width to 350px for breaking - this avoids the circumstance where a two-column layout occurs and the "Course Closed <date>"-type overlay is too large for the tile.
- Added _pinPressHandler - otherwise tabbing off the pin button will always unpin the course (forgot this in the initial version of d2l-course-image-tile)