US94430 Use d2l-css-grid-view-behavior/styles in All Courses
When the US90524 LD flag is on, d2l-all-courses-unified-content will use CSS grids and the new d2l-course-image-tile component like the main widget. Note that because this is only in d2l-all-courses-unified-content, enabling the LD flag will only have an effect when the updated sort order LD flag is also on.