US94430 Fix initial tile sizing
This was previously working, but something snuck in it seems. This new approach is also slightly better - it just uses a couple of CSS variables to define the height of the images rather than each one having its own `resize` event listener. This will end up being a few px off of the previous calculation, since it's not taking the grid-gap into account, but thoroughly meh.