Remove reliance on d2l-tile z-index
d2l-tile has a z-index of 3 set on the menu to allow for things to be in the default slot, use a negative `top` value, and effectively slide in in between the image and menu layers. This was being done to show the various overlays on the course tiles here, but a better solution seems to be just having the overlay within the same slot as the image itself, thereby removing the need to fiddle with z-indexes in d2l-tile.

This fixes a bug (caused by the z-index: 3) where the menu for a different course tile could show on top of another course tile's menu.