Fix long course names
When a course name was sufficiently long, and had no breakable characters (e.g. spaces), it would break the widget as the tile would be the width of the name. Fixed by adding a word-break: break-word, which allows the browser to break the text by word if possible, or by non-breakable characters if necessary.