DE29766 Fix long names with CSS grid
This is a second attempt at this fix - this time verified across browsers. Turns out, `break-word` is not supported anywhere except Chrome and Opera, so the previous fix didn't help all that much.

In fact, the _actual_ issue is that CSS grid items have a default min-width: auto, which means that the tile cannot be smaller than its contents. To fix this, we can set min-width: 0, which allows for them to basically force-shrink themselves even if the course text is long and non-breaking. See https://www.w3.org/TR/css3-grid-layout/#min-size-auto for more information. Defects where you have to read the HTML spec to fix them are the best kind of defects.