US97068 Update sort based on active action
When a set of enrollments is returned, we were previously just updating the local `_enrollmentsSearchAction` from the response. Now, we also want to update the search/filter/sort with the Field values from the returned action. This commit adds this functionality for `sort`, so that an `_enrollmentsSearchAction`'s `sort` field value is applied to the sort menu when it's received.