DE30246 Un-set homepage URL when no access
The situation where this is a problem only comes up in the all courses overlay - in particular, searching for something, then clearing the search, will mean that any course tiles that the user doesn't have access to will not have their URL un-set, meaning the course tile will end up linking to whatever course was in that tile's position. This is a weird Polymer-ism where elements are reused in some ways. To be safe, explicitly set the `_organziationHomepageUrl = null` when there isn't a homepage link on the organization's entity.