Move alerts
Previously, both `d2l-all-courses-segregated-content` and `d2l-all-courses-unified-content` had basically the same implementation of the same thing - being able to handle when a course image fails to be set correctly. While removing the duplicate code is nice, the actual reason for this change is to make adding the tabs to this view easier - it's just easier to have the alerts exist in one place - the over-arching `d2l-all-courses` component - rather than having it in two places, and needing to keep both up to date.