Remove cache bust and extra `embedDepth`
There were cache-busting headers on the request to fetch an organization, but this isn't actually necessary - first, because the request is only cached on the page (via `d2l-fetch`) anyway, so navigating away and then back will reload the organization response. In addition, it's only the image entity we wanted to cache-bust anyway, which leads to...

...not fetching the image right away. Rather than tacking an `embedDepth=1` on to the Link (which is dirty), we should be able to handle both cases: a hydrated image entity, and a linked image entity. It's actually slightly preferable to fetch the linked organization entity, since we don't even _need_ the image initially - we only load it when the course tile scrolls into view, which it may not ever do.

The performance impact of this is immediately noticeable - while it doesn't really speed up initial load much (only really by the amount of time embedding the course image entity took, which wasn't much), it does make opening the All Courses overlay significantly faster, as we're able to use the (likely already fetched) organization and organization image responses, rather than re-fetching them.