US97068 Filter the filter menu!
When a user has their enrollments already grouped by semester/department via the My Courses widget settings, we aren't able to then use the filter menu to further filter them, as it does a logical OR on all the `parentOrganizations` that are sent to the API. When this is the case, we can simply hide those two tabs.

Similarly, if the user has tabs to group their enrollments by RoleAlias, we should hide the Role tab in the filter menu.