DE30396 Use the use-saved-searches attribute
Previously, we were always rendering the tabbed view of both the widget and all courses when `updated-sort-logic` was true. However, there is also the feature flag for the tabbed view itself, `use-saved-searches`. This changes both views to also take this flag's value into account.

In effect, the change here is that when the flag is off, the widget and the overlay both need to kick off their respective `_fetchRoot` calls more manually.