US97924 Improve All Courses tabbed view perf
Previously, all tabs in the All Courses view were actually sharing the same _filteredEnrollments array. This meant we were only ever dealing with a single array, which has some advantages. The downside, however, is that each time the array is rendered, each tab is going to have to update itself to reflect the new state of the array. This resulted in noticeable jank when selecting a tab with a large number of enrollments.

Instead, only the `filtered-enrollments` of the newly-selected tab should be set/updated whenever the user switches tabs. The upside to this is that there is no more UI lag, even with large numbers of enrollments. The downside is that it is possible, if switching to and from tabs very quickly for the first time (i.e. before they're cached), to end up with the wrong tab being selected for the enrollments you're seeing. Fortunately, this can be fixed by just switching to another tab, then back (as at that point, everything has loaded and is cached, so things are speedy-speedy).

This also means we no longer need to delay rendering of the `d2l-all-courses-unified-content` component via the `renderContents` flag, since when it's initially created it'll always be empty anyway.