DE30421 No highlight on inaccessible course
The intended behaviour of a course tile for a course that is inaccessible is that it should not appear to be interactable - that is, it should not get an underline and should not have a "pointer" cursor.

In addition, the entire `d2l-image-tile-image` slot contents should be `aria-hidden`, as this information is already conveyed when the tile is selected, via `_overlayAnnounceText`.

There is still a slight A11Y bug for a disabled link, in that it doesn't properly focus with a keyboard. In particular, the tile is rendering an anchor even if the `href` on the `d2l-image-tile` is null. This is a bug in `d2l-image-tile` itself, though, and should be remedied there (if the `href` is null, the tile should not render the `<a>` within it at all).