US98053 Rework test runner
We've been seeing a lot of test timeouts recently, hence why these tests were disabled in Sauce. In order to get them running again, some concessions had to be made.

In particular, we now only run the tests on the most-relevant components in Sauce (when `updated-sort-logic=true`). This reduces the number of tests we run in each Sauce job to 450, down from 800 or so. We do run all 800 initially, though, in Chrome on Travis.

In addition, we now also divide the Sauce jobs (12 of them) into three groups of four each. This matches our concurrency limit, which means the last four jobs won't time out, since they won't even be started until the first eight are finished.

This also removes `d2l-search-widget-custom-mock` in favour of just mocking out `window.d2lfetch.fetch`, as the former was causing some issues in Firefox.