DE30477 Check returned action for sort
The action we initially get that translates into a tab in the widget doesn't (necessarily) have the sort order set. However, the updated action we receive after fetching the enrollments for said tab-action should have the updated sort on it. This is what we use to determine whether to hide past courses, so by using the _initial_ action rather than the _updated_ action, the `sort` value was always empty, and we were always showing past courses.

This also bumps `d2l-tile`, which had a small related issue, wherein changing sort order would result in inaccessible courses effectively disappearing down to a zero-height course tile.