US98449 Remove updatedSortLogic where not needed
Previously, there were different cases where we'd render what, based on a couple of feature flags. Now, though, we render two almost-entirely-different views based on `updated-sort-config`. If it's on, we always render the `d2l-my-courses-content` and `d2l-all-courses-unified-view` components, both of which use `d2l-course-image-tile`. If it's off, we use the old components. In other words, we only need to know if the flag is on or off at the highest level of the widget, and anything lower down can be entirely ignorant of its existence.