DE30513 Improve "only past courses" alert logic
Previously, we were calling into a function to recalculate whether or not we should be showing the "You don't have any future or past courses" alert. Instead, we can leverage computed properties to make this more resilient to the timing issues that were causing the alert to show when it shouldn't, and not show when it should. In addition, we can greatly simplify the logic, since this alert should never show when updatedSortLogic is false - i.e. it should never show in `d2l-my-courses-content-animated`.