DE30513 Clean up updatedSortLogic usage
Noticed during DE30513 development, but not strictly required for it - it's only really the top-level `d2l-my-courses` component itself, along with `d2l-all-courses` itself, that needs to know what the `updated-sort-logic` value is. For example, `d2l-my-courses-content` is only ever loaded when it's true, and `d2l-my-courses-content-animated` when it's false. This means we can remove a few usage of this value, that do nothing but complicate things.