DE30695 Increase IE11 retries to 5s
This was already increased as part of DE29878, but a case did come up with a client where an overly-full-of-widgets homepage still sees the problem of the tiles overlapping in IE11. Increasing this timeout, and the number of retries, is a bit of a patchwork solution, but ultimately it's a quick fix, doesn't affect other browsers, and won't have any effect on faster homepages in IE11 either.