Use padding and 100% height
Part of a fix to make the enrollment card full-height - that is, when one card has more content (longer name) than others in the same row, all of the cards in the same row should be the same height. There is also a small corresponding CSS change in `d2l-enrollment-card` itself.