DE31001 Hide filter tabs if no search action
In the case where no default department or semester has been set, we don't get an HM Action back from the API to pass on to the associated tab. In this circumstance, we ended up with a blank tab that didn't really do anything; now, we just hide any tabs that don't have a backing action.

This works in the circumstance where both the default department and default semester have been set, when only one is set, and when neither is set. This also doesn't break the name-overriding attributes.