US99396 Add use-enrollment-cards LD flag
This flag will cut between the two possible representations of courses - with the flag on, we use the new (but not yet complete) `d2l-enrollment-card`. When it's off (current default), we use the old `d2l-course-imate-tile`. Once we are ready to release the `d2l-enrollment-card` and for it to entirely replace the tile, we can remove this flag.