DE30960 Add convertUrlToRelative
Let's get this right out of the way: this is awful and nasty, but it's somehow the least-awful-and-nasty. The crux of this issue is that it's possible for the LMS to return URLs that are neither relative, nor canonical, nor on the same origin that we're logged into. Fortunately, the so-far-always-correct case is that when this does happen, we are able to just convert the URL into a relative one, and that will always work.

This is fragile and finicky and not a great practise and _really_ shouldn't be necessary client-side, but it turns out the LMS is bad at knowing what it's own hostname is. So instead, every time we use window.d2lfetch, we can just run the URL through this helper and it'll change it if needed.

I really hope this only exists for a short while, until we can think of a better solution - preferably, the LMS should only be returning URLs that exist on the same origin as the page we're logged into!

Note that this commit is a combination of 6298950e10d0fb10735af52631a1a25a123bcf13 and 39c70fb43b91c61e0fef0424b040a44974c4bcd1, cherry-picked from master.